fix outputlog, diagslog rolling size overflow
(cherry picked from commit 894c2ae86532b792a45373848c9c2f4e913de36e)

+++ lib/ts/Diags.cc
@@ -632,8 +632,8 @@ Diags::should_roll_diagslog()
       if (fstat(fileno(diags_log->m_fp), &buf) != 0)
         return false;
 
-      int size = buf.st_size;
-      if (diagslog_rolling_size != -1 && size >= (diagslog_rolling_size * BYTES_IN_MB)) {
+      off_t size = buf.st_size;
+      if (diagslog_rolling_size != -1 && size >= (static_cast<off_t>(diagslog_rolling_size) * BYTES_IN_MB)) {
         fflush(diags_log->m_fp);
         if (diags_log->roll()) {
           char *oldname = ats_strdup(diags_log->get_name());
@@ -718,8 +718,8 @@ Diags::should_roll_outputlog()
       if (fstat(fileno(stdout_log->m_fp), &buf) != 0)
         return false;
 
-      int size = buf.st_size;
-      if (outputlog_rolling_size != -1 && size >= outputlog_rolling_size * BYTES_IN_MB) {
+      off_t size = buf.st_size;
+      if (outputlog_rolling_size != -1 && size >= static_cast<off_t>(outputlog_rolling_size) * BYTES_IN_MB) {
         // since usually stdout and stderr are the same file on disk, we should just
         // play it safe and just flush both BaseLogFiles
         if (stderr_log->is_init())
